requirements.txt
# TELEGLAS Pro - Python Dependencies
# Production-ready requirements for Python 3.10+

# WebSocket
websockets==12.0
aiohttp==3.9.1

# Telegram Bot
python-telegram-bot==20.7

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database (optional)
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
